updated: refactored util import
- Avoid having a global util import
- Each util function is properly scoped under a semantic folder structure
- The riot internal util will be exported as `dom`, `check`, `tags`, `misc`
