--- lib/browser/common/util.js
@@ -1,4 +1,0 @@
-export * from './util/misc'
-export * from './util/check'
-export * from './util/dom-helpers'
-export * from './util/tag-helpers'
