+++ lib/browser/common/util/dom.js
@@ -2,6 +2,26 @@ import { RE_HTML_ATTRS } from './../global-variables'
 import { isUndefined } from './check'
 
 /**
+ * Shorter and fast way to select multiple nodes in the DOM
+ * @param   { String } selector - DOM selector
+ * @param   { Object } ctx - DOM node where the targets of our search will is located
+ * @returns { Object } dom nodes found
+ */
+export function $$(selector, ctx) {
+  return (ctx || document).querySelectorAll(selector)
+}
+
+/**
+ * Shorter and fast way to select a single node in the DOM
+ * @param   { String } selector - unique dom selector
+ * @param   { Object } ctx - DOM node where the target of our search will is located
+ * @returns { Object } dom node found
+ */
+export function $(selector, ctx) {
+  return (ctx || document).querySelector(selector)
+}
+
+/**
  * Create a generic DOM node
  * @param   { String } name - name of the DOM node we want to create
  * @param   { Boolean } isSvg - should we use a SVG as parent node?
@@ -108,18 +128,4 @@ export function walkNodes(dom, fn, context) {
       dom = next
     }
   }
-}
-
-/**
- * Check whether a DOM node is in stub mode, useful for the riot 'if' directive
- * @param   { Object }  dom - DOM node we want to parse
- * @returns { Boolean } -
- */
-export function isInStub(dom) {
-  while (dom) {
-    if (dom.inStub)
-      return true
-    dom = dom.parentNode
-  }
-  return false
 }
