+++ lib/browser/common/util/misc.js
@@ -1,6 +1,17 @@
 import { isWritable } from './check'
 
 /**
+ * Simple object prototypal inheritance
+ * @param   { Object } parent - parent object
+ * @returns { Object } child instance
+ */
+export function inherit(parent) {
+  function Child() {}
+  Child.prototype = parent
+  return new Child()
+}
+
+/**
  * Specialized function for looping an array-like collection with `each={}`
  * @param   { Array } list - collection of items
  * @param   {Function} fn - callback function
