+++ lib/browser/common/util/tags.js
@@ -1,6 +1,6 @@
 import {
   getAttr
-} from './dom-helpers'
+} from './dom'
 
 import {
   contains,
@@ -143,37 +143,6 @@ export function cleanUpData(data) {
 }
 
 /**
- * Shorter and fast way to select multiple nodes in the DOM
- * @param   { String } selector - DOM selector
- * @param   { Object } ctx - DOM node where the targets of our search will is located
- * @returns { Object } dom nodes found
- */
-export function $$(selector, ctx) {
-  return (ctx || document).querySelectorAll(selector)
-}
-
-/**
- * Shorter and fast way to select a single node in the DOM
- * @param   { String } selector - unique dom selector
- * @param   { Object } ctx - DOM node where the target of our search will is located
- * @returns { Object } dom node found
- */
-export function $(selector, ctx) {
-  return (ctx || document).querySelector(selector)
-}
-
-/**
- * Simple object prototypal inheritance
- * @param   { Object } parent - parent object
- * @returns { Object } child instance
- */
-export function inherit(parent) {
-  function Child() {}
-  Child.prototype = parent
-  return new Child()
-}
-
-/**
  * Set the property of an object for a given key. If something already
  * exists there, then it becomes an array containing both the old and new value.
  * @param { Object } obj - object on which to set the property
@@ -214,6 +183,20 @@ export function arrayishRemove(obj, key, value, ensureArray) {
     else if (obj[key].length == 1 && !ensureArray) obj[key] = obj[key][0]
   } else
     delete obj[key] // otherwise just delete the key
+}
+
+/**
+ * Check whether a DOM node is in stub mode, useful for the riot 'if' directive
+ * @param   { Object }  dom - DOM node we want to parse
+ * @returns { Boolean } -
+ */
+export function isInStub(dom) {
+  while (dom) {
+    if (dom.inStub)
+      return true
+    dom = dom.parentNode
+  }
+  return false
 }
 
 /**
