+++ lib/browser/compiler/index.js
@@ -1,15 +1,9 @@
 import observable from 'riot-observable'
 import compiler from 'riot-compiler'
 import { T_STRING } from './../common/global-variables'
-import {
-  isFunction,
-  isArray,
-  isObject,
-  extend,
-  $$,
-  getAttr,
-  mkEl
-} from './../common/util'
+import { isFunction, isArray, isObject } from './../common/util/check'
+import { extend } from './../common/util/misc'
+import { $$, getAttr, mkEl } from './../common/util/dom'
 
 var
   promise,    // emits the 'ready' event and runs the first callback
