+++ lib/browser/tag/core.js
@@ -12,15 +12,24 @@ import {
   isString,
   isUndefined,
   isObject,
-  isFunction,
+  isFunction
+} from './../common/util/check'
+
+import {
   setAttr,
   getAttr,
-  $$,
+  $$
+} from './../common/util/dom'
+
+import {
   each,
-  mountTo,
-  selectTags,
   extend
-} from './../common/util'
+} from './../common/util/misc'
+
+import {
+  mountTo,
+  selectTags
+} from './../common/util/tags'
 
 /**
  * Export the Tag constructor
