+++ lib/browser/tag/each.js
@@ -1,3 +1,9 @@
+import { isArray } from './../common/util/check'
+import { remAttr, getAttr, getOuterHTML } from './../common/util/dom'
+import { defineProperty, each } from './../common/util/misc'
+import { tmpl } from 'riot-tmpl'
+import Tag from './tag'
+
 import {
   T_STRING,
   T_OBJECT,
@@ -7,23 +13,14 @@ import {
 } from './../common/global-variables'
 
 import {
-  isArray,
-  remAttr,
-  getAttr,
-  getTagName,
+  moveChildTag,
   getTag,
+  getTagName,
   arrayishAdd,
   arrayishRemove,
-  defineProperty,
-  getOuterHTML,
-  moveChildTag,
-  each,
   makeVirtual,
   moveVirtual
-} from './../common/util'
-
-import { tmpl } from 'riot-tmpl'
-import Tag from './tag'
+} from './../common/util/tags'
 
 /**
  * Convert the item looped into an object used to extend the child tag properties
