+++ lib/browser/tag/if.js
@@ -1,4 +1,5 @@
-import { remAttr, unmountAll } from './../common/util'
+import { remAttr } from './../common/util/dom'
+import { unmountAll } from './../common/util/tags'
 import { tmpl } from 'riot-tmpl'
 import { parseExpressions } from './parse'
 import update from './update'
