+++ lib/browser/tag/mkdom.js
@@ -1,15 +1,11 @@
+import { isSVGTag } from './../common/util/check'
+import { mkEl, setInnerHTML, $ } from './../common/util/dom'
+
 import {
   IE_VERSION,
   RE_SPECIAL_TAGS,
   RE_SPECIAL_TAGS_NO_OPTION
 } from './../common/global-variables'
-
-import {
-  isSVGTag,
-  mkEl,
-  setInnerHTML,
-  $
-} from './../common/util'
 
 /*
   Includes hacks needed for the Internet Explorer version 9 and below
