+++ lib/browser/tag/named.js
@@ -1,13 +1,12 @@
 import { tmpl } from 'riot-tmpl'
+import { isBlank } from './../common/util/check'
+import { setAttr, remAttr } from './../common/util/dom'
 
 import {
-  isBlank,
   getImmediateCustomParentTag,
   arrayishAdd,
-  setAttr,
-  remAttr,
   arrayishRemove
-} from './../common/util'
+} from './../common/util/tags'
 
 export default function NamedExpr(dom, attrName, attrValue, parent) {
   this.dom = dom
