+++ lib/browser/tag/parse.js
@@ -3,15 +3,10 @@ import NamedExpr from './named'
 import _each from './each'
 import { tmpl } from 'riot-tmpl'
 import { RIOT_TAG } from './../common/global-variables'
-
-import {
-  isBoolAttr,
-  walkNodes,
-  getAttr,
-  each,
-  getTag,
-  initChildTag
-} from './../common/util'
+import { isBoolAttr } from './../common/util/check'
+import { walkNodes, getAttr } from './../common/util/dom'
+import { each } from './../common/util/misc'
+import { getTag, initChildTag } from './../common/util/tags'
 
 export function parseExpressions(root, tag, expressions, includeRoot) {
   var base = {parent: {children: expressions}}
