+++ lib/browser/tag/styleManager.js
@@ -1,4 +1,4 @@
-import { $, mkEl, setAttr } from './../common/util'
+import { $, mkEl, setAttr } from './../common/util/dom'
 import { WIN } from './../common/global-variables'
 
 var styleNode,
