+++ lib/browser/tag/tag.js
@@ -5,6 +5,33 @@ import update from './update'
 import { tmpl } from 'riot-tmpl'
 import mkdom from './mkdom'
 import { mixin } from './core'
+import { setAttr, remAttr, walkAttrs } from './../common/util/dom'
+
+import {
+  isUndefined,
+  isFunction,
+  isString,
+  isReservedName,
+  isObject,
+  isWritable
+} from './../common/util/check'
+
+import {
+  extend,
+  each,
+  toCamel,
+  contains,
+  defineProperty,
+  inherit
+} from './../common/util/misc'
+
+import {
+  cleanUpData,
+  isInStub,
+  unmountAll,
+  arrayishRemove,
+  getImmediateCustomParentTag
+} from './../common/util/tags'
 
 import {
   GLOBAL_MIXIN,
@@ -16,29 +43,6 @@ import {
   RIOT_TAG_IS,
   RIOT_TAG
 } from './../common/global-variables'
-
-import {
-  isUndefined,
-  isFunction,
-  isString,
-  isReservedName,
-  inherit,
-  cleanUpData,
-  defineProperty,
-  extend,
-  each,
-  walkAttrs,
-  toCamel,
-  isObject,
-  isWritable,
-  isInStub,
-  contains,
-  setAttr,
-  remAttr,
-  unmountAll,
-  arrayishRemove,
-  getImmediateCustomParentTag
-} from '../common/util'
 
 // counter to give a unique id to all the Tag instances
 var __uid = 0
