+++ lib/browser/tag/update.js
@@ -1,5 +1,14 @@
-
 import { tmpl } from 'riot-tmpl'
+import { startsWith, each } from './../common/util/misc'
+import { isFunction, isArray, isWritable } from './../common/util/check'
+import { remAttr, setAttr } from './../common/util/dom'
+
+import {
+  getImmediateCustomParentTag,
+  initChildTag,
+  makeVirtual,
+  arrayishRemove
+} from './../common/util/tags'
 
 import {
   RIOT_PREFIX,
@@ -9,20 +18,6 @@ import {
   __TAG_IMPL,
   IE_VERSION
 } from './../common/global-variables'
-
-import {
-  startsWith,
-  isFunction,
-  isArray,
-  isWritable,
-  each,
-  remAttr,
-  setAttr,
-  getImmediateCustomParentTag,
-  initChildTag,
-  makeVirtual,
-  arrayishRemove
-} from './../common/util'
 
 /**
  * Attach an event to a DOM node
