+++ lib/riot+compiler.js
@@ -1,6 +1,6 @@
 import * as riot from './riot'
 import { compile, parsers } from './browser/compiler/index'
-import { extend } from './browser/common/util'
+import { extend } from './browser/common/util/misc'
 
 function mount(a, b, c) {
   var ret
